Refactor: create Hotel component for displaying hotel details in Dashboard page
