+++ src/pages/Dashboard/Hotel/index.js
@@ -1,15 +1,8 @@
 import * as S from '../../../components/Typography';
 import useHotel from '../../../hooks/api/useHotel';
 import ContentUnavailable from '../../../components/Page/Unavailable';
-import Hotel from '../../../components/Hotel';
 import styled from 'styled-components';
-import { useState } from 'react';
-
-const RoomTypes = {
-  '1': 'Single',
-  '2': 'Single e Double',
-  '3': 'Single, Double e Triple',
-};
+import Hotel from '../../../components/Hotel';
 
 function Page(props) {
   return (
@@ -28,7 +21,7 @@ function Page(props) {
 
 export default function HotelPage() {
   const { hotels, hotelsLoading, hotelsError } = useHotel();
-  const [selected, setSelect] = useState(0);
+  
   if (hotelsLoading) {
     return <Page error>Carregando</Page>;
   }
@@ -36,24 +29,11 @@ export default function HotelPage() {
     return <Page error>{hotelsError.message}</Page>;
   }
 
-  const handleSelected = (id) => {
-    selected === id ? setSelect(0) : setSelect(id);
-  };
-
   return (
     <Page>
       <S.SubtitleTypography variant="h5">Primeiro, escolha seu hotel</S.SubtitleTypography>
       <ContainerHotels>
-        {hotels?.map((hotel) => (
-          <Hotel key={hotel.id} clicked={selected === hotel.id} onClick={() => handleSelected(hotel.id)}>
-            <img src={hotel.image} alt={hotel.name} />
-            <HotelName>{hotel.name}</HotelName>
-            <Subtitle>Tipos de acomodação:</Subtitle>
-            <Detail>{RoomTypes[hotel.capacity]}</Detail>
-            <Subtitle>Vagas disponíveis:</Subtitle>
-            <Detail>{hotel.availableRooms}</Detail>
-          </Hotel>
-        ))}
+        {hotels?.map((hotel) => <Hotel hotel={hotel}/>)}
       </ContainerHotels>
     </Page>
   );
@@ -74,21 +54,4 @@ const ContainerHotels = styled.div`
       flex-direction: column;
       align-items: center;
     }
-`;
-
-const HotelName = styled.span`
-  font-size: 18px;
-  font-weight: bold;
-`;
-
-const Subtitle = styled.p`
-  margin-top: 10px;
-  font-size: 14px;
-  width: 168px;
-  font-weight: bold;
-`;
-
-const Detail = styled(Subtitle)`
-  margin-top: 4px;
-  font-weight: normal;
 `;
